requirements.txt
taipy
pymongo
requests
python-dotenv